+++ run_all_scripts.py
@@ -60,14 +60,14 @@ SCRIPT_CONFIGS = {
         "args": [],
         "time_threshold": timedelta(seconds=90),
     },
-    # "place_taker_orders": {
-    #     "script_name": "place_taker_orders.py",
-    #     "table_id": "latency_experiments.long_running_taker_orders",
-    #     "timestamp_column": "sent_at",
-    #     "filter": "",
-    #     "args": [],
-    #     "time_threshold": timedelta(seconds=180),
-    # },
+    "place_taker_orders": {
+        "script_name": "place_taker_orders.py",
+        "table_id": "latency_experiments.long_running_taker_orders",
+        "timestamp_column": "sent_at",
+        "filter": "",
+        "args": [],
+        "time_threshold": timedelta(seconds=180),
+    },
     "place_stateful_orders": {
         "script_name": "place_stateful_orders.py",
         "table_id": "latency_experiments.long_running_stateful_orders",
